Widen stream priority to i32, ranked onto quiche's urgency (#390)
* Bump -proto

* feat!: widen stream priority to i32, ranked onto quiche's urgency

`SendStream::set_priority` took a `u8`, which forced callers with a
composite ordering (a track priority plus a sequence number, say) to
either rescale or drop resolution before handing it over. Quinn's own
API is an `i32`, so the trait was narrower than the backend underneath
it and the conversion happened in the wrong place.

Widen the trait to `i32` (higher sent first, the `sendOrder`
convention) so callers can bit-pack an ordering into one value. quinn,
noq and iroh now pass it straight through, and qmux's scheduler bands
are keyed by `i32`.

quiche schedules by an 8-bit urgency, so it can't take the value
directly. It now ranks its send streams connection-wide, modeled on
moq's priority queue: streams are grouped into levels by send order,
the top 256 levels get an urgency each, and everything past that shares
the last one. Unlike moq's per-item ranking, levels are per distinct
value, so equal priorities land in one band and quiche round-robins
them, matching quinn and qmux. Rebalancing stops at the first unchanged
overflow band, bounding a change to ~256 level visits rather than the
stream count.

The ranking lives in `DriverState`; the driver drains the urgency
updates it produces, since it is the only holder of the quiche
connection. Updates for streams quiche hasn't created yet are deferred
rather than dropped. Every send stream registers at priority 0 on open,
so a stream that never set one can't outrank a stream that explicitly
asked to go last.

This also fixes an inversion: `web-transport-quiche` previously passed
the trait's priority to quiche's urgency verbatim, where *lower* is
sent first, so a higher priority meant the stream was sent later.


* fix(quiche): ignore priority updates for retired send streams

`Priorities::set` treated an unknown stream id as a new registration, so a
`set_priority` arriving after the driver had already retired the stream
resurrected it. That is reachable with no misuse: a peer STOP_SENDING
retires a stream while the application still holds the `SendStream`.

The resurrected entry is unreachable garbage. Nothing is left to remove it,
since the driver only calls `remove` when it retires a stream from its own
map and that has already happened. It occupies a level, so every live
stream below it is pushed down a band, and enough of them push live streams
into the overflow band. Its urgency update can never be applied either, so
`apply_priorities` defers and re-queues it on every poll forever.

Split registration from re-ranking: `insert` registers, `set` is a no-op
for an id that isn't registered, and both go through a private `rank`.

Reported by Codex on #390.


* fix(quiche): satisfy clippy and rustfmt in the priority test

An orphaned doc comment sat between `PAYLOAD` and `CHECKED`, so clippy
read it as documenting `CHECKED` with a blank line in between and
`empty_line_after_doc_comments` failed the build. It belongs to `PAYLOAD`,
and it referenced a `WINDOW` constant that no longer exists.


* fix(wasm): widen the trait impls to i32

`rs/web-transport-wasm/src/traits.rs` arrived from main in the merge, after
the rest of this branch had already been converted, so its two
`set_priority` impls still took a `u8`. A trait impl has to match the trait
signature exactly, so the wasm32 leg of `just check` could not build.

Forwarding is now direct rather than through `.into()`: the browser's
`sendOrder` is already the same signed range as the trait.


---------

+++ rs/qmux/src/session.rs
@@ -2143,7 +2143,7 @@ pub struct SendStream {
     offset: u64,
     /// Scheduling priority (higher = sent first). Threaded into the queue on
     /// every `push` and relayed to the queue on `set_priority`.
-    priority: u8,
+    priority: i32,
     closed: Option<Error>,
     fin: bool,
 
@@ -2338,7 +2338,7 @@ impl generic::SendStream for SendStream {
     /// Re-prioritization is retroactive: already-queued frames for this stream
     /// move to the new band on the next scheduling decision (the bytes stay put,
     /// preserving per-stream order).
-    fn set_priority(&mut self, order: u8) {
+    fn set_priority(&mut self, order: i32) {
         self.priority = order;
         self.outbound.set_priority(self.id, order);
     }
